Camera Controls
I downloaded a head and beard model from turbosquid. I placed the head into the world. I modified the beard to separate the individual pieces out from the model in maya so that I can remove the indiviudal pieces when the razor collides with the piece. I have also set up the camera controls so the player can rotate around the head and zoom in to get a closer look when holding down the right mouse button. Next I am going to get the razor controls into the game.