Gameplay Loop
I am working on the gameplay loop. I am working on getting it so if you shave an area of the face to much it will cut add a cut texture to the face. Then if there are to many cuts it will kill the customer. I have gotten in the shaving mechanic in and am working on the killing mechanic now.